@@ -9,36 +9,71 @@ import itertools
 import logging
 import warnings
 from collections import Counter
+from collections.abc import Iterable
 from functools import cached_property
-from typing import Any, Callable, Dict, Iterable, List, Optional, Union
+from typing import Any, Callable, Optional, Union
 
 import networkx as nx
 import numpy as np
 import pandas as pd
 
-from ._database_manager import DatabaseManager
-from ._db import DB
-from ._graph_maker import GraphMaker
-from ._the_graph import TheGraph
+from idtrack._database_manager import DatabaseManager
+from idtrack._db import DB
+from idtrack._graph_maker import GraphMaker
+from idtrack._the_graph import TheGraph
 
 
 class Track:
-    """Pathfinding algorithm in prepared bio-ID graph.
+    """Bidirectional path-finding resolver for biological identifiers.
 
-    Uses :py:class:`_the_graph.TheGraph` in order to calculate the matching ID in queried Ensembl release and queried
-    database. It calculates the corresponding IDs from a given ID, by first converting into Ensembl gene ID, then time
-    travelling via connected nodes until requested Ensembl release, finally converting into the requested database.
-    The class uses two important recursive path finding functions: the one is for time travelling and the other is
-    finding the synonymous nodes.
+    `Track` builds and queries a *bio-ID* multigraph that stitches together
+    Ensembl history edges (genes, transcripts, proteins) and cross-reference
+    edges to external databases (UniProt, RefSeq, …).  Given a **source
+    identifier**, a **target Ensembl release**, and/or a **target database**,
+    the class:
+
+    1. **Normalises** the source to an Ensembl *gene* node when necessary.
+    2. **Time-travels** through historical edges—forward or backward—until it
+       reaches the requested release, optionally "beaming-up" through external
+       IDs when the backbone is disconnected.
+    3. **Converts** the resolved Ensembl gene into the requested external
+       database (or returns the gene itself) while annotating the result with
+       confidence scores and the full traversal path.
+
+
+    Two mutually-recursive engines power the search:
+
+    * `_recursive_function`   — depth-first search along temporal edges.
+    * `_recursive_synonymous` — search for synonymous nodes at a single
+      release to enable the external "beam-up".
+
+    Attributes
+    ----------
+    graph : networkx.MultiDiGraph
+        The pre-computed bio-ID graph produced by
+        :py:class:`_graph_maker.GraphMaker`.
+    version_info : dict
+        Metadata about the graph build (Ensembl releases included, build date,
+        Git commit, etc.).
+    _external_entrance_placeholder : dict[bool, int]
+        Sentinel node IDs that mark artificial edges used when an external ID
+        is pulled onto the Ensembl backbone (`False → -1`, `True → 10001`).
+    _external_entrance_placeholders : list[int]
+        Sorted list of the sentinel values above.
     """
 
     # ENSG00000263464
     def __init__(self, db_manager: DatabaseManager, **kwargs):
-        """Class initialization.
+        """Create a `Track` resolver and load (or build) its graph.
 
         Args:
-            db_manager: See parameter in :py:attr:`_graph.Graph.__init__.db_manager`.
-            kwargs: Keyword arguments to be passed to :py:attr:`_the_graph.GraphMaker.get_graph`.
+            db_manager (DatabaseManager): Connection manager that knows how to fetch Ensembl and
+                cross-reference tables from a local cache or a live MySQL mirror.
+                The same instance is forwarded to :py:class:`_graph_maker.GraphMaker`.
+            kwargs: Additional keyword arguments forwarded verbatim to
+                :py:meth:`_graph_maker.GraphMaker.get_graph`.  Common flags include
+                `force_rebuild` (recompute the graph from scratch), `species`
+                (restrict to one taxon), and `cache_dir` (override on-disk cache location).
         """
         self.log = logging.getLogger("track")
         self.db_manager = db_manager
@@ -55,7 +90,7 @@ class Track:
         _the_id: str,
         synonymous_ones: list,
         synonymous_ones_db: list,
-        filter_node_type: set,
+        filter_node_type: set[str],
         the_path: Optional[list] = None,
         the_path_db: Optional[list] = None,
         depth_max: int = 0,
@@ -64,21 +99,48 @@ class Track:
     ):
         """Helper method to be used in :py:meth:`_graph.Track.synonymous_nodes`.
 
-        It recursively looks at the graphs and returns a synonymous IDs under two main constrains: the depth
-        and the final node type. The depth is simply defined as the total max number of certain database jump to
-        be included in the final path. The recursive search does not walk across two nodes that has the same node type
-        to prevent this method to travel through time.
+        Recursively explore the bio-ID graph to collect **synonymous paths**
+        starting at `_the_id` and ending on a node whose *type* is a member of `filter_node_type`.
+
+        A *path* is a list of node identifiers (`_the_path`) together with a
+        parallel list of their node-type strings (`_the_path_db`).
+        The search is **breadth-limited**: the *depth* of a path is defined as
+        the **maximum count of any single node-type** it contains
+        (e.g. a path with three `external` nodes has depth 3).  Recursion
+        stops when that depth would exceed `depth_max`.
+
+        Additional pruning rules:
+
+        - The walk never visits the same node twice (no cycles).
+        - It never traverses two consecutive edges whose source and target
+            share the **same node-type**—this prevents "time-travel" within the
+            Ensembl history backbone.
+        - When `ensembl_backbone_shallow_search` is *True*, the search is
+            restricted to the **reverse** direction except for node-types listed
+            in :py:attr:`DB.nts_bidirectional_synonymous_search`.
+
+        On reaching a terminating node the method *appends* the discovered
+        paths to `synonymous_ones` and `synonymous_ones_db`. It does not
+        return anything. Results are accumulated in `synonymous_ones` and `synonymous_ones_db`.
 
         Args:
-            _the_id: The name of the node such as Ensembl ID or external ID.
-            synonymous_ones: Todo.
-            synonymous_ones_db: Todo.
-            the_path: Todo.
-            the_path_db: Todo.
-            depth_max: Todo.
-            filter_node_type: Todo.
-            from_release: Todo.
-            ensembl_backbone_shallow_search: Todo
+            _the_id (str): Identifier of the starting node (Ensembl or
+                external).
+            synonymous_ones (list): Mutable list that will receive
+                each successful identifier path.
+            synonymous_ones_db (list): Mutable list that will receive
+                the corresponding node-type paths.
+            filter_node_type (set[str]): Allowed node-types for the **final**
+                node of a path (e.g. `{'ensembl_gene'}`).
+            the_path (list | None): Current path leading to
+                `_the_id`; *None* for the root invocation.
+            the_path_db (list | None): Node-type counterpart of
+                `the_path`; *None* for the root invocation.
+            depth_max (int): Maximum allowed depth as defined above.
+            from_release (int | None): If given, only keep terminal
+                nodes that are *active* in this Ensembl release.
+            ensembl_backbone_shallow_search (bool): Activate the
+                shallow, mostly-reverse search mode described above.
         """
 
         def decide_terminate_externals():
@@ -179,24 +241,39 @@ class Track:
         self,
         the_id: str,
         depth_max: int,
-        filter_node_type: set,
+        filter_node_type: set[str],
         from_release: Optional[int] = None,
         ensembl_backbone_shallow_search: bool = False,
     ):
-        """Todo.
+        """Public wrapper around :py:meth:`_recursive_synonymous`.
+
+        The method returns **all minimal-length synonym paths** emanating from `the_id`.
+
+        The function first runs a *default* depth search determined by
+        `DB.external_search_settings['synonymous_max_depth']`.
+        If no synonym is found and `depth_max` is greater than that default,
+        a second, deeper search is attempted.
+
+        For every distinct *target* node the shortest path is kept; longer
+        paths to the same target are discarded.
 
         Args:
-            the_id: Todo.
-            depth_max: Todo.
-            filter_node_type: Todo.
-            from_release: Todo.
-            ensembl_backbone_shallow_search: Todo. assemly bridge'i bozar cunku direction iki tarafli olmali
+            the_id (str): Source identifier.
+            depth_max (int): Maximum search depth to try if the default search fails.
+            filter_node_type (set[str]): Node-types that are acceptable for the
+                *target* node(s).  Must *not* include the generic
+                `'external'` type—specify the concrete external DB instead.
+            from_release (int | None): Constrain targets to those active in this Ensembl release.
+            ensembl_backbone_shallow_search (bool): If *True*,
+                restricts the graph traversal as explained in :py:meth:`_recursive_synonymous`.
 
         Returns:
-            Todo.
+            list[list[list[str]]]: A list whose elements are `[identifier_path, node_type_path]` pairs,
+                each representing the minimal route to one synonymous node.
 
         Raises:
-            ValueError: Todo.
+            ValueError: If `filter_node_type` improperly contains the generic
+                external type, or if `depth_max` is incompatible with `ensembl_backbone_shallow_search`.
         """
         if DB.nts_external in filter_node_type:
             raise ValueError(f"Define which external database: `{filter_node_type}`.")
@@ -258,18 +335,39 @@ class Track:
 
     @staticmethod
     def get_from_release_and_reverse_vars(lor: list, p: int, mode: str):
-        """Todo.
+        """Derive a list of `(release, reverse)` tuples.
+
+        Derive a list of `(release, reverse)` tuples that indicate *which*
+        Ensembl release to start the graph walk from and *whether* that walk
+        should move **backwards in time**.
+
+        Given a collection of active-range intervals `lor` and a pivot
+        release `p`, the algorithm selects one or two release points per
+        interval depending on `mode`:
+
+        * 'closest' - choose the release **nearest** to `p` within or at
+            the ends of the interval.
+        * 'distant' - choose the release **farthest** from `p` within the
+            interval.
+
+        The boolean in each tuple is *True* when the walk should start **after**
+        the selected release and move backwards (i.e. "reverse mode"), and
+        *False* when it should move forwards.
 
         Args:
-            lor: Todo.
-            p: Todo.
-            mode: Todo.
+            lor (list): List of inclusive
+                `(first_release, last_release)` intervals in ascending order.
+            p (int): Pivot release around which "closest" or "distant" is
+                evaluated.
+            mode (str): Either `'closest'` or `'distant'`.
 
         Returns:
-            Todo.
+            list[tuple[int, bool]]: Release / reverse-flag pairs, ordered in
+            the sequence they should be tried by the path-finder.
 
         Raises:
-            ValueError: Todo.
+            ValueError: If an interval in `lor` is malformed, `mode` is
+                not recognised, or internal consistency checks fail.
         """
         result = list()
         # Note that there is no need to consider assembly here because always this function is used in the context of
@@ -309,24 +407,57 @@ class Track:
     def _choose_relevant_synonym_helper(
         self, from_id, synonym_ids, to_release: int, from_release: Optional[int], mode: str
     ):
-        """Todo.
+        """Select the **most temporally relevant** synonym(s) for an Ensembl gene-ID family.
+
+        The method evaluates each candidate in `synonym_ids` against the
+        *target* release `to_release` and, when applicable, the *source*
+        release `from_release`.  Its job is to decide **where** the path
+        should *enter* the Ensembl backbone and **whether** the remainder of
+        the traversal must run in *reverse* (new → old) order.
+
+        Selection strategy
+        ------------------
+        1. **Fixed `from_release`** - If the caller already knows the release
+           of the starting node, every candidate is paired with that same
+           release and the correct *reverse* flag is derived trivially.
+        2. **Non-backbone start** - When the starting node is not an
+           Ensembl-gene backbone ID, the synonym whose **active range edge** is
+           *closest* (or *farthest*, per `mode`) to `to_release` is chosen.
+        3. **Backbone start** - If the query is itself an Ensembl-gene, the
+           algorithm first looks for *overlapping* ranges between the query
+           and each synonym; if none overlap, it falls back to the distance
+           rule described in step 2.
 
         Args:
-            from_id: Todo.
-            synonym_ids: Todo.
-            to_release: Todo.
-            from_release: Todo.
-            mode: Todo.
+            from_id (str): Identifier from which the path search will start.
+            synonym_ids (Sequence[str]): Ensembl IDs considered synonyms of
+                `from_id` (typically the *same* gene with different version
+                numbers).
+            to_release (int): Target Ensembl release that the overall conversion
+                aims for.
+            from_release (int | None): Release in which `from_id` is known to
+                be active.  If *None*, the method infers a suitable release for
+                each candidate.
+            mode (str): Either `'closest'` or `'distant'`—controls whether
+                the synonym chosen should minimise or maximise its distance to
+                `to_release`.
 
         Returns:
-            Todo.
+            list[list[Union[str, int, bool]]]: One or more triplets of the form
+            `[synonym_id, entry_release, reverse]` where:
+
+            * `synonym_id` - the chosen synonym,
+            * `entry_release` - release at which to join the backbone, and
+            * `reverse` - *True* if the subsequent history walk must run
+              backwards in time.
 
         Raises:
-            ValueError: Todo.
+            ValueError: If no synonym satisfies the distance/overlap criteria
+                or if `mode` is invalid.
         """
-        # synonym_ids should be ensembl of the same id (different versions)
         distance_to_target = list()
         candidate_ranges = list()
+        # synonym_ids should be ensembl of the same id (different versions)
 
         if from_release is not None:
             candidate_ranges.extend([[i, from_release, from_release > to_release] for i in synonym_ids])
@@ -400,24 +531,46 @@ class Track:
         # given from release
 
     def choose_relevant_synonym(
-        self, the_id: str, depth_max: int, to_release: int, filter_node_type: set, from_release: Optional[int]
+        self, the_id: str, depth_max: int, to_release: int, filter_node_type: set[str], from_release: Optional[int]
     ):
-        """Todo.
+        """Wrapper that **discovers, clusters, and ranks** synonymous Ensembl candidates for a given identifier.
+
+        The function performs three steps:
+
+        1. **Discover** paths to *all* Ensembl-gene nodes that share the same
+           biological identity (`synonymous_nodes`).
+        2. **Cluster** those paths by **gene ID** (ignoring version).
+        3. **Rank** each cluster with
+           :py:meth:`_choose_relevant_synonym_helper`, selecting the entry release
+           (and direction) that best suits `to_release`.
 
         Args:
-            the_id: Todo.
-            depth_max: Todo.
-            to_release: Todo.
-            filter_node_type: Todo.
-            from_release: Todo
+            the_id (str): Source identifier (Ensembl or external).
+            depth_max (int): Maximum depth passed to
+                :py:meth:`synonymous_nodes`; governs how far the synonym search is
+                allowed to roam through external nodes.
+            to_release (int): Target Ensembl release required by the overall
+                conversion.
+            filter_node_type (set[str]): Node-types that the synonym search
+                must terminate on (usually `{'ensembl_gene'}`).
+            from_release (int | None): Known active release of `the_id`.  If
+                *None*, the helper will infer one.
 
         Returns:
-            Todo.
-        """
-        # help to choose z for a->x->z3,6,9
+            list[list[Any]]: A list whose elements are
 
-        # filter_node_type == 'ensembl_gene'
+            `[synonym_id, entry_release, reverse, identifier_path, node_type_path]`
+
+            where the last two items reproduce the path returned by
+            :py:meth:`synonymous_nodes`.
+
+        Notes:
+            *The method purposefully keeps **all** equally-ranked candidates;
+            further tie-breaking is deferred to the main path-scoring routine.*
+        """
         syn = self.synonymous_nodes(the_id, depth_max, filter_node_type, from_release=from_release)
+        # help to choose z for a->x->z3,6,9
+        # filter_node_type == 'ensembl_gene'
         # it also returns itself, which is important
 
         syn_ids: dict = dict()
@@ -450,19 +603,33 @@ class Track:
         return result  # new_from_id, new_from_rel, new_reverse, path, path_db
 
     def get_next_edges(self, from_id: str, from_release: int, reverse: bool, debugging: bool = False):
-        """Todo.
+        """Enumerate **chronologically admissible** history edges from a node.
+
+        Starting at `from_id` and release `from_release`, the method scans
+        outgoing (or incoming, when `reverse` is *True*) edges whose
+        timestamps allow the path to **advance** in the desired temporal
+        direction.  It collapses duplicate "same-ID" transitions and flags
+        self-loops so that later heuristics can treat branch points and tips
+        differently.
 
         Args:
-            from_id: Todo.
-            from_release: Todo.
-            reverse: Todo.
-            debugging: Todo.
+            from_id (str): Current node from which the search will step.
+            from_release (int): Release at which the current node is known to
+                exist.
+            reverse (bool): *False* to walk **forward** in history
+                (old → new), *True* to walk **backward** (new → old).
+            debugging (bool): If set, disables the duplicate-edge
+                collapse so that unit tests can inspect the raw edge set.
 
         Returns:
-            Todo.
+            list[list[Union[int, bool, str, int]]]: Sorted list of edge
+            descriptors, each of which is
+
+            `[edge_release, is_self_loop, src_node, dst_node, multiedge_key]`.
 
         Raises:
-            ValueError: Todo.
+            ValueError: If inconsistent multi-edges (same nodes, same release)
+                are detected—this signals a corrupted graph build.
         """
         edges: list = list()
         more_than_one_edges: dict = dict()
@@ -516,17 +683,40 @@ class Track:
         return sorted(edges, reverse=reverse)  # sort based on history
 
     def should_graph_reversed(self, from_id, to_release):
-        """Todo.
+        """Determine the **temporal orientation** of the graph walk.
+
+        Given an identifier that is active in one or more release intervals,
+        the routine decides whether the subsequent path-finder must move
+        **forward in time**, **backward in time**, or explore *both* directions
+        in parallel in order to reach the target release.
+
+        The decision is based on the *closest* boundary of every active
+        interval returned by
+        :py:meth:`Track.get_from_release_and_reverse_vars` (`mode='closest'`).
 
         Args:
-            from_id: Todo.
-            to_release: Todo.
+            from_id (str): The starting identifier (Ensembl gene, transcript,
+                protein, or external ID).
+            to_release (int): The Ensembl release the user wishes to convert
+                *to*.
 
         Returns:
-            Todo.
+            tuple[str, Union[int, tuple[int, int]]]:
+
+                * `'forward'` - walk old → new, starting at the **earliest**
+                    release in which *from_id* is active
+                    &nbsp;&nbsp;&nbsp;→ return `('forward', start_release)`
+                * `'reverse'` - walk new → old, starting at the **latest**
+                    active release
+                    &nbsp;&nbsp;&nbsp;→ return `('reverse', start_release)`
+                * `'both'` - split search: one forward walk and one reverse
+                    walk
+                    &nbsp;&nbsp;&nbsp;→ return `('both', (forward_start,
+                    reverse_start))`
 
         Raises:
-            ValueError: Todo.
+            ValueError: If *from_id* is never active in or around
+                `to_release` (i.e. no viable starting release can be found).
         """
         n = self.graph.get_active_ranges_of_id[from_id]  # calculates ensembl_gene nodes and also any other node.
         m = Track.get_from_release_and_reverse_vars(n, to_release, mode="closest")
@@ -556,19 +746,50 @@ class Track:
         external_jump: Optional[float] = None,
         multiple_ensembl_transition: bool = False,
     ) -> set:
-        """Todo.
+        """Enumerate every admissible **history path** from *from_id* at *from_release* to *to_release*.
+
+        The algorithm performs a depth-first traversal of the Ensembl history
+        edges.  Whenever it becomes stranded on a non-backbone node it may
+        "beam-up" via a *synonym path* through an external database, subject to
+        the constraints in `external_settings`:
+
+        * `synonymous_max_depth` - maximum depth of a synonym search.
+        * `jump_limit` - maximum number of external "beam-up" jumps allowed.
+        * `nts_backbone` - canonical node-type of the Ensembl backbone.
+
+        Additional flags control the initial conditions:
+
+        * Setting `external_jump` to *np.inf* **disables** external jumps.
+            Setting it to *None* **enables** them with the counter reset to `0`.
+        * `multiple_ensembl_transition` allows the algorithm to time-travel to
+            a *different* release while still on an external node; this is useful
+            when *from_release* was inferred and might not actually connect.
 
         Args:
-            from_id: Todo.
-            from_release: Todo.
-            to_release: Todo.
-            reverse: Todo.
-            external_settings: Todo.
-            external_jump: Todo.
-            multiple_ensembl_transition: Todo.
+            from_id (str): Identifier to start the search from.
+            from_release (int): Release number where *from_id* is considered
+                active.
+            to_release (int): Target Ensembl release.
+            reverse (bool): If *True*, traverse the graph **backwards** in time;
+                otherwise forwards.
+            external_settings (dict): Copy of
+                :py:attr:`DB.external_search_settings` that governs depth,
+                jump limits, and backbone node-type.
+            external_jump (float | None): Current external-jump count
+                (*None* starts from zero, *np.inf* forbids any jump).
+            multiple_ensembl_transition (bool): Permit the synonym
+                engine to select a *different* release for an external node when
+                no path exists at *from_release*.
 
         Returns:
-            _description_
+            set[tuple[tuple[str, str, int]]]:
+                A set of **edge-lists**.  Each edge is stored as
+                `(src, dst, key)`; an empty walk that terminates immediately
+                is represented by `((None, from_id, None),)`.
+
+        Notes:
+            *The method is intentionally side-effect free; it constructs all
+            intermediate data on the stack and returns a fresh set.*
         """
 
         def _recursive_function(
@@ -776,20 +997,47 @@ class Track:
         increase_jump_until: int = 0,
         from_release_inferred: bool = False,
     ) -> tuple:
-        """Todo.
+        """Run :py:meth:`path_search` under **progressively relaxed settings**.
+
+        Run :py:meth:`path_search` under **progressively relaxed settings** until at
+        least one viable path is found—or every relaxation level is exhausted.
+
+        Four search stages are attempted in order:
+
+        1. **Backbone-only** - external jumps disabled.
+        2. **External enabled** - allow external jumps; increment synonym depth
+            and jump limit after each failure up to
+            `increase_depth_until`/`increase_jump_until`.
+        3. **Backbone with multiple-Ensembl transition** - external disabled but
+            permit starting release to shift on external nodes.
+        4. **External + multiple-transition** - most permissive search, with
+            iterative depth/jump relaxation as in stage 2.
 
         Args:
-            from_id: Todo.
-            from_release: Todo.
-            to_release: Todo.
-            reverse: Todo.
-            go_external: Todo.
-            increase_depth_until: Todo.
-            increase_jump_until: Todo.
-            from_release_inferred: Todo.
+            from_id (str): Identifier to convert.
+            from_release (int): Release at which the search begins.
+            to_release (int): Desired target release.
+            reverse (bool): Traverse the Ensembl history backwards if *True*,
+                forwards otherwise.
+            go_external (bool): If *False*, skip any stage that
+                requires external jumps.
+            increase_depth_until (int): Additional synonym-search
+                depth to allow beyond the default.
+            increase_jump_until (int): Additional external-jump count
+                to allow beyond the default.
+            from_release_inferred (bool): *Reserved for future use.*
+                Indicates that *from_release* was chosen automatically rather
+                than provided by the user.
 
         Returns:
-            Todo.
+            tuple[tuple[tuple[str, str, int]]]:
+                All paths discovered by the most restrictive stage that yielded
+                **at least one** result, returned as an immutable tuple.
+
+        Notes:
+            The function copies and mutates
+            :py:attr:`DB.external_search_settings` internally; the caller’s
+            copy is not modified.
         """
         es: dict = copy.deepcopy(DB.external_search_settings)
         idu = increase_depth_until + es["synonymous_max_depth"]
@@ -862,6 +1110,24 @@ class Track:
 
     @cached_property
     def _calculate_node_scores_helper(self):
+        """Build and cache helper look-ups for node-scoring.
+
+        The property constructs two complementary data structures:
+
+        * **filter_set** - the union of
+            (1) every external-database node-type present in the graph, and (2) every Ensembl-specific
+            node-type (gene, transcript, translation, …) across *all* assemblies.
+            This set can therefore be passed unmodified to
+            :py:meth:`synonymous_nodes` to ask for "anything that is not an assembly-less backbone gene".
+        * **ensembl_include** - a mapping
+            `{form → set(node_type_str)}` where each value lists the node-types
+            that should be considered equivalent to that *form* (e.g. *gene*, *transcript*, *translation*) when
+            computing richness metrics.
+
+        Returns:
+            tuple[set[str], dict[str, set[str]]]:
+                `(filter_set, ensembl_include)` exactly as described above.
+        """
         form_list = [i for i in self.graph.available_forms if i != DB.backbone_form]
         ensembl_include = dict()
         filter_set = copy.deepcopy(self.graph.available_external_databases)
@@ -872,17 +1138,28 @@ class Track:
         return filter_set, ensembl_include
 
     def calculate_node_scores(self, the_id, ens_release) -> list:
-        """A metric to choose from multiple targets.
+        """Rank competing Ensembl targets by the "richness" of their synonyms.
+
+        The method counts, within a radius of *two* synonym hops, how many unique
+        identifiers of various categories point to each candidate and returns the
+        counts as **negative integers** so that *smaller* is *better* for the
+        up-stream sorter.
 
         Args:
-            the_id: Todo.
-            ens_release: Todo.
+            the_id (str): Identifier that is being converted.
+            ens_release (int): Target Ensembl release; only synonyms active in this
+                release are considered.
 
         Returns:
-            Todo.
+            list: `[-ext, -form₁, -form₂]` where
+                * `ext`  - number of distinct **external-database** synonyms.
+                * `form₁` - number of distinct synonyms of the most important Ensembl form (typically *gene*).
+                * `form₂` - number of distinct synonyms of the second form (typically *transcript* or *translation*).
 
         Raises:
-            ValueError: Todo.
+            ValueError: If the graph does not expose exactly the two expected
+                non-backbone forms, or if a synonym node's type cannot be mapped to
+                *external*, *form₁*, or *form₂*.
         """
         form_list = [i for i in self.graph.available_forms if i != DB.backbone_form]
         form_importance_order = ["gene", "transcript", "translation"]
@@ -921,29 +1198,54 @@ class Track:
         all_possible_paths: tuple,
         reduction: Callable,
         remove_na: str,
-        from_releases: Iterable,
+        from_releases: Iterable[int],
         to_release: int,
         score_of_the_queried_item: float,
         return_path: bool,
         from_id: str,
     ) -> dict:
-        """Todo.
+        """Collapse a set of candidate paths into the single best path per target.
+
+        For each path produced by the search engine the function:
+
+        1. Computes an **edge-score aggregate** using `reduction` while handling
+        missing values as directed by `remove_na`.
+        2. Tallies *external* statistics (steps, jumps, initial conversion
+        confidence) and *assembly* statistics (number of priority drops, final
+        priority).
+        3. Packs all metrics into a dictionary and stores it under the key of the
+        path's final destination node.
+        4. Keeps only the lexicographically "smallest" dictionary per destination
+        via :py:meth:`_path_score_sorter_single_target`.
 
         Args:
-            all_possible_paths: Todo.
-            reduction: Todo.
-            remove_na: Todo.
-            from_releases: Todo.
-            to_release: Todo.
-            score_of_the_queried_item: Todo.
-            return_path: Todo.
-            from_id: Todo.
+            all_possible_paths (tuple): Sequence of edge-lists representing
+                every admissible walk returned by the path-finder.
+            reduction (Callable): Function such as
+                `np.mean` or `sum` used to collapse edge weights into one number.
+            remove_na (str): How to treat *NaN* edge weights - one of
+                `'omit'`, `'to_1'`, `'to_0'`.
+            from_releases (Iterable[int]): Release that each path starts from;
+                must align with `all_possible_paths`.
+            to_release (int): Target release - needed to know whether an edge is
+                traversed forward or reverse.
+            score_of_the_queried_item (float): Fallback weight for the implicit edge
+                that represents the *query ID* itself.
+            return_path (bool): If *True*, embed the full edge-list inside each
+                score dict under the key `'the_path'`.
+            from_id (str): Original identifier being converted - echoed back in the
+                score dict for traceability.
 
         Returns:
-            Todo.
+            dict: Mapping
+            `{destination_id → best_score_dict}`.
+            Each *score dict* contains (inter alia)
+            `assembly_jump`, `external_jump`, `external_step`,
+            `edge_scores_reduced`, and `ensembl_step`.
 
         Raises:
-            ValueError: Todo.
+            ValueError: If an unexpected edge encoding is encountered, if an edge
+                score is invalid/∞, or if `remove_na` is set to an unknown mode.
         """
 
         def er_maker_for_initial_conversion(n1, n2, n3, n4):
@@ -1017,10 +1319,12 @@ class Track:
                 assembly_jump, current_priority = 0, max(step_pri)
             else:
                 the_path = tuple(
-                    tuple(list(the_step[:3]) + [er_maker_for_initial_conversion(*the_step)])
-                    if len(the_step) == 4 and the_step[-1] in self._external_entrance_placeholders
-                    # -1, 10001 is when you have external to graph conversion and from_release is None.
-                    else the_step
+                    (
+                        tuple(list(the_step[:3]) + [er_maker_for_initial_conversion(*the_step)])
+                        if len(the_step) == 4 and the_step[-1] in self._external_entrance_placeholders
+                        # -1, 10001 is when you have external to graph conversion and from_release is None.
+                        else the_step
+                    )
                     for the_step in the_path
                 )
                 assembly_jump, step_pri, current_priority = self.minimum_assembly_jumps(the_path)
@@ -1046,15 +1350,24 @@ class Track:
         return max_score  # dict of dict
 
     def edge_key_orientor(self, n1: str, n2: str, n3: int):
-        """Todo.
+        """Return the stored orientation of a multigraph edge.
+
+        For multigraphs every logical edge is stored *once*, but the caller may
+        hold `(u, v, k)` or `(v, u, k)`.
+        This helper resolves the ambiguity so that subsequent attribute look-ups
+        succeed.
 
         Args:
-            n1: Todo.
-            n2: Todo.
-            n3: Todo.
+            n1 (str): One endpoint of the edge.
+            n2 (str): The other endpoint.
+            n3 (int): Edge key (index) within the `networkx` multi-edge.
 
         Returns:
-            Todo.
+            tuple[str, str, int]: A triple that is guaranteed to exist as written
+            in `self.graph`.
+
+        Raises:
+            AssertionError: If neither orientation is present in the graph.
         """
         if self.graph.has_edge(n1, n2, n3):
             edge_key = (n1, n2, n3)
@@ -1065,19 +1378,28 @@ class Track:
         return edge_key
 
     def path_step_possible_assembly_jumps(self, n1, n2, n3, n4=None):
-        """Todo.
+        """Return the genome assemblies that can legally be used for a single edge.
+
+        The helper inspects the edge that connects `n1` → `n2` and filters the
+        assemblies recorded on that edge against the *release* constraint `n4`:
+
+        * **None** - the edge is treated as backbone history; the result is the
+            graph-wide default assembly (usually the build on which the backbone was
+            constructed).
+        * **int** - keep only assemblies whose *release set* contains that single release.
+        * **set[int]** - keep assemblies whose release set intersects the provided set.
 
         Args:
-            n1: Todo.
-            n2: Todo.
-            n3: Todo.
-            n4: Todo.
-
-        Raises:
-            ValueError: Todo.
+            n1 (str): Source node identifier.
+            n2 (str): Destination node identifier.
+            n3 (int): Edge key within the NetworkX multigraph.
+            n4 (int | set[int] | None, optional): Release filter as described above.
 
         Returns:
-            Todo.
+            list[str]: Sorted list of assembly names (e.g. `["GRCh37", "GRCh38"]`).
+
+        Raises:
+            ValueError: If `n4` is of an unsupported type.
         """
         edge_key = self.edge_key_orientor(n1, n2, n3)
         if n4 is None:
@@ -1092,15 +1414,28 @@ class Track:
             raise ValueError
 
     def minimum_assembly_jumps(self, the_path, step_pri=None, current_priority=None) -> tuple:
-        """Todo.
+        """Compute the penalty incurred by assembly downgrades along a path.
+
+        Each path step may be annotated with one or more candidate assemblies.
+        These are translated into priority values via
+        `DB.assembly_mysqlport_priority`.  The algorithm walks the path,
+        tracking the current priority and counting how many times it must drop
+        to a lower priority value—each drop constitutes an "assembly jump"
+        penalty.
 
         Args:
-            the_path: Todo.
-            step_pri: Todo.
-            current_priority: Todo.
+            the_path (Iterable[tuple]): Sequence of edge descriptors; each element is
+                either `(n1, n2, k)` or `(n1, n2, k, release)`.
+            step_pri (list[int] | None, optional): Priority list for the first edge.
+                If None, it is derived from `the_path`.
+            current_priority (int | None, optional): Starting priority.  If None,
+                initialised to `max(step_pri)`.
 
         Returns:
-            Todo.
+            tuple[int, list[int], int]:
+                - `assembly_jump` - total number of priority drops.
+                - `step_pri` - priority list of the last processed edge.
+                - `current_priority` - priority value after the final edge.
         """
         assemblies = [self.path_step_possible_assembly_jumps(*i) for i in the_path]
         # should be sorted for bisect function to work properly in '_minimum_assembly_jumps_helper'
@@ -1117,18 +1452,25 @@ class Track:
 
     @staticmethod
     def _minimum_assembly_jumps_helper(step_pri, current_priority, priorities):
-        """Todo.
+        """Internal worker for :py:meth:`minimum_assembly_jumps`.
+
+        Given the priority sets for the remaining edges, iterate until all have
+        been consumed while updating the *current* assembly priority and counting
+        how often it must drop.
 
         Args:
-            step_pri: Todo.
-            current_priority: Todo.
-            priorities: Todo.
-
-        Raises:
-            ValueError: Todo.
+            step_pri (list[int]): Priority values of the edge currently under
+                consideration.
+            current_priority (int): Priority value inherited from previous steps.
+            priorities (list[list[int]]): Priority lists for the *rest* of the path,
+                **already sorted** for correct bisecting.
 
         Returns:
-            Todo.
+            tuple[int, list[int], int]: Same three-tuple as documented in
+            :py:meth:`minimum_assembly_jumps`.
+
+        Raises:
+            ValueError: If the priority lattice enters an impossible state.
         """
 
         def next_priority(p):
@@ -1160,17 +1502,26 @@ class Track:
         return penalty, step_pri, current_priority
 
     @staticmethod
-    def _path_score_sorter_single_target(lst_of_dict: list) -> dict:
-        """Todo.
+    def _path_score_sorter_single_target(lst_of_dict: list[dict]) -> dict:
+        """Select the best score dictionary for *one* conversion target.
+
+        The input is a list of dictionaries produced by
+        :py:meth:`calculate_score_and_select`.  Each dictionary is converted into a
+        tuple according to the *lexicographic importance order*
+
+        `("assembly_jump", "external_jump", "external_step",
+        "edge_scores_reduced", "ensembl_step")`
+
+        and the dictionary with the **smallest** tuple is returned.
 
         Args:
-            lst_of_dict: Todo.
-
-        Raises:
-            ValueError: Todo.
+            lst_of_dict (list[dict]): Candidate score dictionaries for this target.
 
         Returns:
-            Todo.
+            dict: The chosen "winner" score dictionary.
+
+        Raises:
+            ValueError: If the input list is empty.
         """
         importance_order = (  # the variables from to_add in 'calculate_score_and_select'.
             "assembly_jump",
@@ -1189,18 +1540,39 @@ class Track:
         return lst_of_dict[best_score_index]  # choose the best & shortest
 
     def _path_score_sorter_all_targets(self, dict_of_dict: dict, from_id: str, to_release: int) -> dict:
-        """Todo.
+        """Select the overall best target(s).
+
+        Select the **overall best target(s)** once every candidate Ensembl node
+        has itself been reduced to its single best path.
+
+        The method linearises several per-path metrics into an *importance
+        order* (see the tuple at the top of the function), then:
+
+        1. Computes that ordered score for **each** pair `(ensembl_gene, final_target)`.
+        2. Finds the global minimum; if multiple pairs tie:
+            * Prefer the target whose identifier is identical to `from_id`.
+            * If more than one Ensembl gene still tie, fall back on
+                :py:meth:`calculate_node_scores` to favour the "richer" node.
+        3. Returns a *pruned* copy of `dict_of_dict` that contains only the
+           surviving Ensembl genes, each with only the winning `final_elements` entry.
+           Additional provenance is written to `filter_scores`.
 
         Args:
-            dict_of_dict: Todo.
-            from_id: Todo.
-            to_release: Todo.
-
-        Raises:
-            ValueError: Todo.
+            dict_of_dict (dict): Nested result of
+                :py:meth:`calculate_score_and_select`.  Keys are candidate
+                Ensembl genes; values are dictionaries that already contain
+                *one* best path per final target.
+            from_id (str): Original query identifier; used to break ties in
+                favour of "same as input".
+            to_release (int): Target Ensembl release; forwarded to
+                :py:meth:`calculate_node_scores` during tie-breaking.
 
         Returns:
-            Todo.
+            dict: A reduced version of `dict_of_dict` holding only the winner(s) and enriched with a
+                `final_elements[*]['filter_scores']` sub-dict that records the filters applied.
+
+        Raises:
+            ValueError: If `dict_of_dict` is empty.
         """
         importance_order = (
             # the variables from to_add in 'calculate_score_and_select'.
@@ -1223,7 +1595,7 @@ class Track:
         if not len(dict_of_dict) > 0:
             raise ValueError
 
-        minimum_scores: Dict[tuple, list] = dict()
+        minimum_scores: dict[tuple, list] = dict()
 
         for dct in dict_of_dict:
             for target in dict_of_dict[dct]["final_conversion"]["final_elements"]:
@@ -1295,27 +1667,59 @@ class Track:
         deprioritize_lrg_genes: bool = True,
         return_ensembl_alternative: bool = True,
     ):
-        """Todo.
+        """End-to-end ID conversion workflow.
+
+        Starting from `from_id` the routine
+
+        1. Determines the correct **time-travel direction** if
+           `from_release` is unspecified.
+        2. Enumerates *all* admissible paths with
+           :py:meth:`get_possible_paths` (forward and/or reverse).
+        3. Collapses those paths with :py:meth:`calculate_score_and_select`.
+        4. Optionally converts the surviving Ensembl gene(s) into
+           `final_database` via :py:meth:`_final_conversion`.
+        5. Optionally applies a final global selection with
+           :py:meth:`_path_score_sorter_all_targets`.
+
+        The output structure mirrors this decision tree and, when
+        `return_path` is *True*, embeds the full edge list so that callers
+        can audit every hop.
 
         Args:
-            from_id: Todo.
-            from_release: Todo.
-            to_release: Todo.
-            final_database: Todo.
-            reduction: Todo.
-            remove_na: Todo.
-            score_of_the_queried_item: Todo.
-            go_external: Todo.
-            prioritize_to_one_filter: Todo.
-            return_path: Todo.
-            deprioritize_lrg_genes: Todo.
-            return_ensembl_alternative: Todo.
+            from_id (str): Source identifier (Ensembl, UniProt, RefSeq, …).
+            from_release (int | None): Starting Ensembl release.
+                *None* → infer from the graph.
+            to_release (int | None): Target Ensembl release.
+                Defaults to the newest release contained in the graph.
+            final_database (str | None): External database to
+                convert into.  *None* → stay on the Ensembl gene.
+            reduction (Callable): Function (e.g. `numpy.mean`)
+                used to collapse per-edge weights.  Must accept an iterable of
+                floats and return a float.
+            remove_na (str): Strategy for *NaN* edge weights -
+                `'omit'`, `'to_1'`, or `'to_0'`.
+            score_of_the_queried_item (float): Weight assigned to
+                the implicit edge that represents `from_id` itself.
+            go_external (bool): Allow jumps through external
+                databases when the backbone is disconnected.
+            prioritize_to_one_filter (bool): After all scoring,
+                keep only the single globally best target.
+            return_path (bool): Embed the full edge list(s) in the
+                returned dictionary.
+            deprioritize_lrg_genes (bool): If *True* and other
+                results exist, drop LRG_* genomic regions from the final set.
+            return_ensembl_alternative (bool): When converting to an
+                external database, also return the Ensembl gene as a fallback.
 
         Returns:
-            Todo.
+            dict | None:
+                * **dict** - Structured result as described above.
+                * **None** - No admissible path was found.
 
         Raises:
-            ValueError: Todo.
+            ValueError: For non-callable `reduction`, unsupported
+                `remove_na` modes, unknown `final_database` values, or
+                logical inconsistencies detected during processing.
         """
         if not callable(reduction):
             raise ValueError
@@ -1419,6 +1823,24 @@ class Track:
                 return self._path_score_sorter_all_targets(converted, from_id, to_release)
 
     def _create_priority_list_ensembl(self, from_id: str, to_release: int):
+        """Build a priority list of assemblies in which `from_id` is active.
+
+        The priorities are the numeric **assembly rankings** defined in
+        :py:attr:`DB.assembly_mysqlport_priority` (smaller numbers mean
+        higher priority).
+
+        Args:
+            from_id (str): Ensembl gene identifier.
+            to_release (int): Target Ensembl release; only assemblies that
+                contain this release are considered.
+
+        Returns:
+            list[int]: Sorted list of priority values (ascending).
+
+        Raises:
+            ValueError: If the gene is not active in *any* assembly at
+                `to_release`.
+        """
         ceg = self.graph.combined_edges_genes[from_id]
         ceg_assembly_list = sorted({j for i in ceg for j in ceg[i] if to_release in ceg[i][j]})
         if len(ceg_assembly_list) == 0:
@@ -1429,25 +1851,40 @@ class Track:
     def _final_conversion_dict_prepare(
         confidence: Union[int, float],
         sysns: list,
-        paths: Optional[List[List]],
+        paths: Optional[list[list]],
         min_priority_list: list,
         len_priority_list: list,
         add_ass_jump_list: list,
         final_database: str,
     ):
-        """Todo.
+        """Assemble the final-conversion section that will be attached to a candidate path.
+
+        The section contains a global conversion-confidence flag plus one
+        entry per synonym that survived the path-finding stage.  When
+        `paths` is None the structure is identical but omits the
+        `'the_path'` member to save memory.
 
         Args:
-            confidence: Todo.
-            sysns: Todo.
-            paths: Todo.
-            min_priority_list: Todo.
-            len_priority_list: Todo
-            add_ass_jump_list: Todo.
-            final_database: Todo.
+            confidence (int | float): Heuristic confidence for the *whole*
+                conversion step - `0` for "perfect", larger values for
+                fallback scenarios, `np.inf` when no conversion was
+                possible.
+            sysns (list): List of synonym identifiers *in the same order*
+                as the metric lists below.
+            paths (list[list] | None): One walk (edge list) per synonym, or
+                *None* if the caller does not want to expose paths.
+            min_priority_list (list): Minimum assembly priority reached
+                by each walk.
+            len_priority_list (list): Number of distinct assembly
+                priorities encountered by each walk.
+            add_ass_jump_list (list): Additional assembly-jump penalty
+                incurred during the synonym hop itself.
+            final_database (str): Name of the database these synonyms belong
+                to (e.g. `'uniprot'` or `DB.nts_ensembl["gene"]`).
 
         Returns:
-            Todo.
+            dict: Nested dictionary ready to be stored under the key
+            `'final_conversion'`.
         """
         if paths is not None:
             return {
@@ -1486,30 +1923,58 @@ class Track:
         return_path: bool,
         return_ensembl_alternative: bool,
     ):
-        """Todo.
+        """Convert an Ensembl gene node to the requested external database.
+
+        Convert an Ensembl gene node to the requested external database and
+        merge the result back into `converted`.
+
+        The routine:
+
+        1. Builds every legal synonym path from `cnvt` to
+           `final_database` that is active in `ens_release` (or in any
+           release as a fallback).
+        2. Computes assembly-jump penalties for each path.
+        3. Calls :py:meth:`_final_conversion_dict_prepare` to create the
+           conversion sub-dict.
+        4. Optionally falls back to returning the Ensembl gene itself when no
+           synonym exists and `return_ensembl_alternative` is True.
 
         Args:
-            converted: Todo.
-            cnvt: Todo.
-            final_database: Todo.
-            ens_release: Todo.
-            return_path: Todo.
-            return_ensembl_alternative: Todo.
+            converted (dict): The current accumulator being built by
+                :py:meth:`convert`.
+            cnvt (str): Ensembl gene identifier that is undergoing final
+                conversion.
+            final_database (str): Target external database.
+            ens_release (int): Target Ensembl release.
+            return_path (bool): If True, embed the path(s) that lead to each
+                synonym.
+            return_ensembl_alternative (bool): When no synonym can be found,
+                add a fallback entry that keeps the Ensembl gene.
 
         Returns:
-            Todo.
+            dict: The same `converted` dict, updated in place (and also
+            returned for convenience).
         """
 
         def _final_conversion_path(gene_id: str, target_db: str, from_release: int):
-            """Todo.
+            """Produce every synonym path from an Ensembl gene to *one* external database.
+
+            The search is attempted twice:
+            1. Release-restricted - only synonyms active in
+               `from_release`; confidence = 0.
+            2. Release-agnostic - synonyms active in any release;
+               confidence = 1.
 
             Args:
-                gene_id: Todo.
-                target_db: Todo.
-                from_release: Todo.
+                gene_id (str): Ensembl gene identifier acting as the source.
+                target_db (str): Desired external database.
+                from_release (int): Ensembl release that paths must reach
+                    (first attempt).
 
             Returns:
-                Todo.
+                tuple[list[list], int]:
+                    - `syn_ids_path` - list of edge-lists, one per synonym.
+                    - `confidence`   - 0 for strict, 1 for fallback search.
             """
 
             def _final_conversion_path_helper(res_syn_mth, er: Optional[int]):
@@ -1556,16 +2021,18 @@ class Track:
                 confidence = 1
                 return the_paths_no_ens_rel, confidence
 
-        def _final_conversion_helper(conv_dict, conv_dict_key, a_path):
-            """Todo.
+        def _final_conversion_helper(conv_dict: dict, conv_dict_key: str, a_path: list):
+            """Re-compute assembly-jump metrics after appending the external conversion path.
 
             Args:
-                conv_dict: Todo.
-                conv_dict_key: Todo.
-                a_path: Todo.
+                conv_dict (dict): Parent entry in `converted` holding the
+                    pre-conversion metrics.
+                conv_dict_key (str): Key of that entry (the Ensembl gene).
+                a_path (list): Edge list representing one complete walk
+                    (history + synonym hop).
 
             Returns:
-                Todo.
+                list[int | float]: `[assembly_jump, min_priority, n_priorities]`.
             """
             # the path will continue with final conversion so assembly penalty should be calculated again.
             _s1, _s2 = conv_dict[conv_dict_key]["final_assembly_priority"]
@@ -1605,20 +2072,28 @@ class Track:
 
         return converted
 
-    def identify_source(self, dataset_ids: list, mode: str):
-        """Todo.
+    def identify_source(self, dataset_ids: list[str], mode: str):
+        """Infer the most likely origin (assembly and/or Ensembl release) of a heterogeneous identifier list.
+
+        The function tallies how often each origin triple appears among
+        `dataset_ids` and returns the counts sorted in descending order.
 
         Args:
-            dataset_ids: Todo.
-            mode: Todo.
+            dataset_ids (list[str]): Collection of identifiers to analyse.
+            mode (str): Granularity of the origin to extract - one of
+                - 'complete'                  → (assembly, db, release)
+                - 'ensembl_release'           → release only
+                - 'assembly'                  → assembly only
+                - 'assembly_ensembl_release'  → (assembly, release)
 
         Returns:
-            Todo.
+            list[tuple[Any, int]]: Pairs `(origin, count)` sorted by
+            frequency.
 
         Raises:
-            ValueError: Todo.
+            ValueError: If `mode` is not one of the recognised values.
         """
-        possible_trios: List[Any] = list()
+        possible_trios: list[Any] = list()
         for di in dataset_ids:
             # assembly = self.get_external_assembly()
             if mode == "complete":
@@ -1635,9 +2110,13 @@ class Track:
         return list(Counter(possible_trios).most_common())
 
     def convert_optimized_multiple(self):
-        """Accept multiple ID list and return the most optimal set of IDs, minimizing the clashes.
+        """Placeholder for a **batch-optimised** converter.
+
+        The intended behaviour is to accept *multiple* query IDs and choose a
+        conversion target for each such that cross-sample clashes (e.g.
+        duplicate loci) are minimised.
 
         Raises:
-            NotImplementedError: Todo.
+            NotImplementedError: Always - the optimisation strategy is not yet implemented.
         """
         raise NotImplementedError  # TODO
