@@ -1,8 +1,10 @@
 """Nox sessions."""
+
 import os
 import shlex
 import shutil
 import sys
+import tempfile
 from pathlib import Path
 from textwrap import dedent
 
@@ -17,7 +19,7 @@ except ImportError:
     sys.exit(1)
 
 package = "idtrack"
-python_versions = ["3.8", "3.9"]
+python_versions = ["3.9", "3.10", "3.11", "3.12"]
 nox.options.sessions = (
     "pre-commit",
     "safety",
@@ -123,12 +125,26 @@ def precommit(session: Session) -> None:
         activate_virtualenv_in_precommit_hooks(session)
 
 
-@session(python=python_versions)
+@nox.session(python=python_versions)
 def safety(session: Session) -> None:
     """Scan dependencies for insecure packages."""
-    requirements = session.poetry.export_requirements()
-    session.install("safety")
-    session.run("safety", "check", "--full-report", f"--file={requirements}")
+    # Create a temporary file for exported requirements
+    with tempfile.NamedTemporaryFile(mode="w+", delete=False, suffix=".txt") as reqs_file:
+        reqs_path = Path(reqs_file.name)
+
+    try:
+        # Export the requirements from poetry
+        session.run("poetry", "export", "--format=requirements.txt", "--without-hashes", "--output", str(reqs_path))
+
+        # Install safety 2.x
+        session.install("safety>=2,<3")
+
+        # Run safety check
+        session.run("safety", "check", "--full-report", f"--file={reqs_path}")
+    finally:
+        # Clean up the temp file
+        if reqs_path.exists():
+            reqs_path.unlink()
 
 
 @session(python=python_versions)
@@ -137,7 +153,15 @@ def mypy(session: Session) -> None:
     args = session.posargs or ["idtrack", "tests", "docs/conf.py"]
     session.install(".")
     session.install(
-        "mypy", "pytest", "types-pkg-resources", "types-requests", "types-attrs", "types-PyMySQL", "types-PyYAML"
+        # types-pkg-resources, which was unceremoniously disappeared from the
+        # Python supply chain. instead `types-setuptools`
+        "mypy",
+        "pytest",
+        "types-setuptools",
+        "types-requests",
+        "types-attrs",
+        "types-PyMySQL",
+        "types-PyYAML",
     )
     session.run("mypy", *args)
 
